requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.51.2
faiss-cpu==1.8.0
sentence-transformers==3.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.2
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.0
numpy==1.26.4
tiktoken==0.7.0
python-dotenv==1.0.0
requests==2.31.0
BeautifulSoup4==4.12.2
aiofiles==23.2.0
jinja2==3.1.2
